feat: serve assets in unit tests
When unit-testing a directive, and the template of the directive contains, for example

    <img src="/assets/angular.png" />

a 404 warning is displayed in the karma console because the image isn't served by Karma.

This commit configures karma to also serve assets, to avoid that warning (and possible
other complications caused by the absence of assets).